% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df4.Plot.by.Factr.R
\name{df4.Plot.by.Factr}
\alias{df4.Plot.by.Factr}
\title{Creating Dataset for Plot.by.Factr}
\usage{
df4.Plot.by.Factr(var, df)
}
\arguments{
\item{var}{Vector of factor variables.}

\item{df}{Dataset.}
}
\value{
The output from  \code{\link{df4.Plot.by.Factr}}
}
\description{
Creating Dataset for Plot.by.Factr
}
\examples{
df <- sample_data[c("Formal","Informal","L.Both","No.Loan",
"sex","educ","political.afl","married",
"havejob","rural","age","Income","Networth","Liquid.Assets",
"NW.HE","fin.knowldge","fin.intermdiaries")]
CN = colnames(df)
var <- c("educ", "rural", "sex", "havejob", "political.afl")
df4.Plot.by.Factr(var,df)

}
