% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.by.Factr.R
\name{Plot.by.Factr}
\alias{Plot.by.Factr}
\title{Plot by Factor and Continuous}
\usage{
Plot.by.Factr(XXX)
}
\arguments{
\item{XXX}{object to be plotted.}
}
\value{
The output from  \code{\link{Plot.by.Factr}}.
}
\description{
Plot by Factor and Continuous
}
\examples{
\dontrun{
df <- sample_data[c("Formal","Informal","L.Both","No.Loan",
"sex","educ","political.afl","married",
"havejob","rural","age","Income","Networth","Liquid.Assets",
"NW.HE","fin.knowldge","fin.intermdiaries")]
CN = colnames(df)
var <- c("educ","rural","sex","havejob","political.afl")
df4.Plot.by.Factr(var,df)
Plot.by.Factr(df4.Plot.by.Factr(var,df)$Summ.Stats.long)

}
}
