% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvot.by.Factr.R
\name{Pvot.by.Factr}
\alias{Pvot.by.Factr}
\title{Pivot Table by Factor}
\usage{
Pvot.by.Factr(df)
}
\arguments{
\item{df}{The data frame of factor variables.}
}
\value{
The output from  \code{\link{Pvot.by.Factr}}.
}
\description{
Pivot Table by Factor
}
\examples{
df <- sample_data[c("multi.level",
"Formal","L.Both","No.Loan",
"region", "sex", "educ", "political.afl",
"married", "havejob", "rural",
"fin.knowldge", "fin.intermdiaries")]
Pvot.by.Factr(df)
}
