% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.by.Factr.R
\name{Stats.by.Factr}
\alias{Stats.by.Factr}
\title{Summary Statistics by Factor}
\usage{
Stats.by.Factr(var, df)
}
\arguments{
\item{var}{The vector to set summary statistics.}

\item{df}{The name of the Data set.}
}
\value{
The output from  \code{\link{Stats.by.Factr}}.
}
\description{
Summary Statistics by Factor
}
\examples{
\dontrun{
df <- sample_data[c("Formal","Informal","L.Both","No.Loan",
"sex","educ","political.afl","married",
"havejob","rural","age","Income","Networth","Liquid.Assets",
"NW.HE","fin.knowldge","fin.intermdiaries")]
CN = colnames(df)
var <- c("educ","rural","sex","havejob","political.afl")
Stats.by.Factr(var, df)
}

}
