% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.by.Factr.R
\name{Plot.by.Factr}
\alias{Plot.by.Factr}
\title{Plot by Factor}
\usage{
Plot.by.Factr(XXX, name.levels)
}
\arguments{
\item{XXX}{object to be plotted.}

\item{name.levels}{name object.}
}
\value{
The output from  \code{\link{Plot.by.Factr}}.
}
\description{
Plot by Factor
}
\examples{
df <- sample_data[c("Formal","Informal","L.Both",
"No.Loan", "sex","educ","political.afl","married",
"havejob","rural","age","Income","Networth","Liquid.Assets",
"NW.HE","fin.knowldge","fin.intermdiaries")]
CN = colnames(df)
var <- c("educ","rural")

name.levels <- c("Formal","Informal","L.Both","No.Loan",
"sex","educ","political.afl","married",
"havejob","rural","age","Income","Networth","Liquid.Assets",
"NW.HE","fin.knowldge","fin.intermdiaries")

XXX <- df4.Plot.by.Factr(var,df)$Summ.Stats.long
Plot.by.Factr(XXX, name.levels)
}
