% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pvot.by.Factr.R
\name{Pvot.by.Factr}
\alias{Pvot.by.Factr}
\title{Pivot Table by Factor}
\usage{
Pvot.by.Factr(dsc)
}
\arguments{
\item{dsc}{The matrix of factor variables.}
}
\value{
The output from  \code{\link{Pvot.by.Factr}}.
}
\description{
Pivot Table by Factor
}
