% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.by.Factr.R
\name{Stats.by.Factr}
\alias{Stats.by.Factr}
\title{Summary Statistics by Factor}
\usage{
Stats.by.Factr(var, df)
}
\arguments{
\item{var}{The vector to set summary statistics.}

\item{df}{The name of the Data set.}
}
\value{
The output from  \code{\link{Stats.by.Factr}}.
}
\description{
Summary Statistics by Factor
}
\examples{
\dontrun{
var <- c("cyl", "vs", "am")
Stats.by.Factr(var, mtcars)
}
}
